Extract the remaining tests
Should be good to test the github action now
