+++ inducer/graph/graph_node.py
@@ -6,7 +6,6 @@
 @note: contains a function to create the networkx graphs
 '''
 import logging
-import unittest
 
 
 class GraphNode():
@@ -60,36 +59,3 @@ class GraphNode():
 
     def __hash__(self):
         return hash(self.index)
-
-
-class TestGraphNode(unittest.TestCase):
-
-    def setUp(self):
-        GraphNode.index = 0
-        GraphNode.count = 0
-
-    def testInit(self):
-        GraphNode()
-        GraphNode(color=1, index=1)
-        self.assertEqual(GraphNode.index, 1)
-        self.assertEqual(GraphNode.count, 2)
-
-    def testComparison(self):
-        n1 = GraphNode()
-        n2 = GraphNode()
-        self.assertEqual(n1 < n2, True)
-        self.assertEqual(n1 <= n2, True)
-        self.assertEqual(n1 == n2, False)
-        self.assertEqual(n1 > n2, False)
-        self.assertEqual(n1 >= n2, False)
-        self.assertEqual(n1 != n2, True)
-        self.assertEqual(n2 == GraphNode(index=1), True)
-
-    def testHash(self):
-        self.assertEqual(hash(GraphNode()), 0)
-        self.assertEqual(hash(GraphNode()), 1)
-
-
-if __name__ == "__main__":
-    # import sys;sys.argv = ['', 'Test.testName']
-    unittest.main()
