+++ inducer/strong_stable_set.py
@@ -13,18 +13,11 @@ Email:   lama3790@mylaurier.ca, fras2560@mylaurier.ca
 Version: 2015-10-21
 -------------------------------------------------------
 """
-
-from inducer.helper import make_clique, make_diamond, make_cycle
-import networkx as nx
 from networkx import find_cliques, graph_clique_number
 from networkx import complement
-import unittest
 
 
 def strong_stable_set(G):
-    '''
-
-    '''
     clique = graph_clique_number(G)
     result = None
     for stable in stable_set(G):
@@ -41,35 +34,3 @@ def stable_set(G):
     co_g = complement(G)
     for clique in find_cliques(co_g):
         yield clique
-
-
-class Test(unittest.TestCase):
-
-    def setUp(self):
-        pass
-
-    def tearDown(self):
-        pass
-
-    def testStableSet(self):
-        expect = [[0], [1], [2, 3]]
-        for i in stable_set(make_diamond()):
-            self.assertEqual(i in expect, True)
-
-    def testStrongStableSet(self):
-        g = nx.Graph()
-        g.add_node(0)
-        # check a triangle
-        result = strong_stable_set(make_clique(3))
-        self.assertEqual(result.nodes(), g.nodes())
-        # check a diamond
-        result = strong_stable_set(make_diamond())
-        self.assertEqual(result.nodes(), g.nodes())
-        # C5
-        result = strong_stable_set(make_cycle(5))
-        self.assertEqual(result, None)
-
-
-if __name__ == "__main__":
-    # import sys;sys.argv = ['', 'Test.testName']
-    unittest.main()
