--- inducer/test.py
@@ -1,25 +1,0 @@
-"""
--------------------------------------------------------
-testing module
-holds some Test Cases
--------------------------------------------------------
-Author:  Dallas Fraser
-ID:      110242560
-Email:   fras2560@mylaurier.ca
-Version: 2014-09-11
--------------------------------------------------------
-"""
-
-
-import unittest
-
-
-class Tester(unittest.TestCase):
-
-    def setUp(self):
-        # called when starting test
-        pass
-
-    def tearDown(self):
-        # called when done test
-        pass
